requirements.txt
streamlit
streamlit-lottie
youtube-transcript-api
python_dotenv
google-generativeai